Adjusted Book Details for Author
Needed to change Book Details so that it would find the author if he exists in DB and add him to this book, and also add the book to the authors booksWritten. Also made a new author if didnt exist when adding the book. Also made the if statements a bit shorter for errors